+++ app/views/authenticated/index.html.erb
@@ -3,5 +3,5 @@
   <p>Find me in app/views/authenticated/index.html.erb</p>
   <p><%= link_to "Back to welcome page", root_path, class: tw_standard_link %></p>
   <p><%= link_to "Manage your account", edit_user_registration_path, class: tw_standard_link %></p>
-  <p><%= button_to "Log out", destroy_user_session_path, method: :delete, class: tw_cancel_button %></p>
+  <p><%= button_to "Log out", destroy_user_session_path, method: :delete, class: "mt-1 mb-2 #{tw_cancel_button}" %></p>
 </div>
