+++ app/views/devise/confirmations/new.html.erb
@@ -10,7 +10,7 @@
     </div>
 
     <div class="actions">
-      <%= f.submit "Resend confirmation instructions", class: tw_success_button %>
+      <%= f.submit "Resend confirmation instructions", class: "mt-1 mb-2 #{tw_success_button}" %>
     </div>
   <% end %>
