+++ app/views/devise/passwords/edit.html.erb
@@ -19,7 +19,7 @@
     </div>
 
     <div class="actions">
-      <%= f.submit "Change my password", class: tw_success_button %>
+      <%= f.submit "Change my password", class: "mt-1 mb-2 #{tw_success_button}" %>
     </div>
   <% end %>
