+++ app/views/devise/passwords/new.html.erb
@@ -10,7 +10,7 @@
     </div>
 
     <div class="actions">
-      <%= f.submit "Send me reset password instructions", class: tw_success_button %>
+      <%= f.submit "Send me reset password instructions", class: "mt-1 mb-2 #{tw_success_button}" %>
     </div>
   <% end %>
