+++ app/views/devise/registrations/edit.html.erb
@@ -33,13 +33,13 @@
     </div>
 
     <div class="actions">
-      <%= f.submit "Update", class: tw_success_button %>
+      <%= f.submit "Update", class: "mt-1 mb-2 #{tw_success_button}" %>
     </div>
   <% end %>
 
   <h3>Cancel my account</h3>
 
-  <p>Unhappy? <%= button_to "Cancel my account", registration_path(resource_name), form: { data: { turbo_confirm: "Are you sure?" } }, method: :delete, class: tw_alert_button %></p>
+  <p>Unhappy? <%= button_to "Cancel my account", registration_path(resource_name), form: { data: { turbo_confirm: "Are you sure?" } }, method: :delete, class: "mt-1 mb-2 #{tw_alert_button}" %></p>
 
   <%= link_to "Back", :back, class: tw_standard_link %>
 </div>
