+++ app/views/devise/registrations/new.html.erb
@@ -23,7 +23,7 @@
     </div>
 
     <div class="actions">
-      <%= f.submit "Sign up", class: tw_success_button %>
+      <%= f.submit "Sign up", class: "mt-1 mb-2 #{tw_success_button}" %>
     </div>
   <% end %>
