+++ app/views/devise/sessions/new.html.erb
@@ -20,7 +20,7 @@
     <% end %>
 
     <div class="actions">
-      <%= f.submit "Log in", class: tw_success_button %>
+      <%= f.submit "Log in", class: "mt-1 mb-2 #{tw_success_button}" %>
     </div>
   <% end %>
