+++ app/views/devise/unlocks/new.html.erb
@@ -10,7 +10,7 @@
     </div>
 
     <div class="actions">
-      <%= f.submit "Resend unlock instructions", class: tw_success_button %>
+      <%= f.submit "Resend unlock instructions", class: "mt-1 mb-2 #{tw_success_button}" %>
     </div>
   <% end %>
